Improve dependency resolution process to not use `FileCollection`
`extendsFrom` superconfigurations are not resolved when the extending configurations are resolved.
This change at least improves how the configurations and dependencies are managed by providing more proper dependency targets.
The extending configuraitons cannot really be queried either since it requires a resolution for them to have the dependencies we want.

see: https://discuss.gradle.org/t/resolving-a-configuration-does-not-resolve-the-parent-configurations-that-it-extendsfrom/24413/2
see: https://discuss.gradle.org/t/configuration-and-transitive-dependency-management-by-extension/23824
closes https://github.com/mkobit/jenkins-pipeline-shared-libraries-gradle-plugin/issues/2