Generated newline character in loglines is platform independent
This makes it so the code generation is not platform dependent.
Previously, the code generated would be line dependent on the system it was built on because the newline character was inserted at code generation time.
Now, it is inserted based on the runtime context.

related to https://github.com/mkobit/jenkins-pipeline-shared-libraries-gradle-plugin/issues/31