Ability to use `@NonCPS` in library code
This may introduce extra time for first compilation since it requires the dependencies for integration test to be resolved rather than delaying it until the integration test compilation is required.
This implementation uses the transitive version instead of some mapped or hardcoded dependency.
It also unfortunately exposes all of the other types to the compilation classpath which may not actually be available at Jenkins runtime.
Can get feedback on this as it is used, and come back to the implementation later.

closes https://github.com/mkobit/jenkins-pipeline-shared-libraries-gradle-plugin/issues/17