Add in classpath resource generation to handle changes with latest Jenkins LTS and JEP 200
The newest Jenkins LTS 2.107.1 added in more security around serialization/deserialization.
This appears to cause some issues with the generated source retriever as it is serialized and stored with XStream (apparently?).
The generated classes (right now only LocalLibraryRetriever) that are stored to disk need to be updated to handle this new requirement.
Based on what I saw at https://jenkins.io/blog/2018/03/15/jep-200-lts/ and https://jenkins.io/blog/2018/01/13/jep-200/, I added a resource generation task to include in the whitelist.
This seems to fix the error logs but I'll have to wait and see if that is an actual fix.

closes https://github.com/mkobit/jenkins-pipeline-shared-libraries-gradle-plugin/issues/61