chore(ci): add GitHub Actions CI and Publish Workflows (#121)
## Description

This PR only includes the GitHub Actions workflows without removing the
existing CircleCI setup.

Also, this CI currently runs only on 'ubuntu-latest' runners.

I wanted to make as few source code changes as possible. Given that some
dependencies were exclusively in JCenter, which has been shut down, I
had to make some dependency changes.

The idea is for this CI to act as checks and become the first step in
revamping this project. This will allow merging the changes at
https://github.com/mkobit/jenkins-pipeline-shared-libraries-gradle-plugin/pull/119
with more confidence.

### Gradle

Upgraded Gradle to 6.8

- https://docs.gradle.org/6.8/release-notes.html

```sh
./gradlew wrapper --gradle-version=6.8
```

### Other

I also included some properties to `.editorconfig` to avoid wildcard
imports. These are only applicable to IntelliJ.

### Issues

Work towards:

-
https://github.com/mkobit/jenkins-pipeline-shared-libraries-gradle-plugin/issues/110

### Test Evidence

I opened a PR in my own fork space to validate this workflow run:
https://github.com/albertocavalcante/fork-jenkins-pipeline-shared-libraries-gradle-plugin/pull/1/checks?check_run_id=42265945090

Publish failed due to the missing credentials to publish to Maven
Central. Nevertheless, it shouldn't be a priority for now as there's
still a long way to go before the next release.